+++ CHANGELOG.md
@@ -27,7 +27,7 @@
     + NSMutableArray methods: `-removeFirstObject`, `-shiftFirstObject`, `-popLastObject`, `-reverseObjects`, `-shuffleObjects`
     + NSOrderedSet methods: `objectOrNilAtIndex:`
     + `-[NSTimeZone hoursFromGMT]`
-    + UIApplication methods: `-registerForRemoteNotificationsWithCompletion:`
+    + UIApplication methods: `-appWindow`, `-registerForRemoteNotificationsWithCompletion:`, `-appIconFile`, `-appIconImage`
     + UIDevice methods: `-modelName`, `-deviceToken`, `-deviceTokenString`
     + `ESNetworkHelper` class
     + `-[UIView snapshotViewAfterScreenUpdates:]`